multiclient: background retry dial clients
