+++ multiclient/map.go
@@ -48,6 +48,15 @@ func (m *Map) Set(key string, value *rpc.Client) {
 	m.m[key] = value
 }
 
+func (m *Map) Replace(key string, value *rpc.Client) {
+	m.lock.Lock()
+	defer m.lock.Unlock()
+
+	if _, ok := m.m[key]; ok {
+		m.m[key] = value
+	}
+}
+
 func (m *Map) Get(key string) *rpc.Client {
 	m.lock.RLock()
 	defer m.lock.RUnlock()
@@ -99,6 +108,19 @@ func (m *Map) Keys() []string {
 	for k := range m.m {
 		urls[index] = k
 		index++
+	}
+	return urls
+}
+
+func (m *Map) NilClients() []string {
+	m.lock.RLock()
+	defer m.lock.RUnlock()
+
+	urls := make([]string, 0)
+	for k, v := range m.m {
+		if v == nil {
+			urls = append(urls, k)
+		}
 	}
 	return urls
 }
